[API CHANGE] Drop css::accessibility::XAccessibleStateSet
which is internal API, unused (as far as I can tell) by external
users.
This state is purely a bitset
(as implemented by utl::AccessibleStateSetHelper)
so we can just return it as a 64-bit value. We could probably
get away with a 32-bit value, but it seems wise to leave some
space for future needs.

This shaves significant time off the performance profiles
of code that loads very complex shapes, because this state
is frequently used, and we no longer need to allocate a return
value on the heap for every call.

+++ qadevOOo/tests/java/mod/_svtools/AccessibleIconChoiceCtrl.java
@@ -213,7 +213,7 @@ public class AccessibleIconChoiceCtrl extends TestCase {
             "com.sun.star.accessibility.AccessibleStateType.MANAGES_DESCENDANTS",
         };
 
-        short[] expectedStates = new short[] {
+        long[] expectedStates = new long[] {
             com.sun.star.accessibility.AccessibleStateType.ENABLED,
             com.sun.star.accessibility.AccessibleStateType.FOCUSABLE,
             com.sun.star.accessibility.AccessibleStateType.SHOWING,
